Queue issueue

+++ Icris.uservicebus/Icris.uServiceBus.Core/Queues/FileSystemMessage.cs
@@ -10,7 +10,7 @@ using System.Transactions;
 
 namespace Icris.uServiceBus.Core.Queues
 {
-    
+
 
     public class FileSystemMessage<T>
     {
@@ -18,32 +18,63 @@ namespace Icris.uServiceBus.Core.Queues
 
         SimpleFileLock fileLock;
         string path;
-        public FileSystemMessage(string path)
+        int timeout;
+        DateTime expiration = DateTime.Now;
+        T content;
+        public FileSystemMessage(string path, int timeout)
         {
-            lock(creationlock)
+            try
             {
+                this.timeout = timeout;
                 this.path = path;
-                this.fileLock = SimpleFileLock.Create(path + ".lock", TimeSpan.FromMinutes(10));
-                if (!fileLock.TryAcquireLock())
-                {
-                    throw new Exception("Unable to aqcuire lock");
-                }                
+                Lock();
+                content = JsonConvert.DeserializeObject<T>(File.ReadAllText(path));
+            }
+            catch (Exception e)
+            {
+                throw new Exception("Unable to lock the message");
             }
         }
+        private void Lock()
+        {
+            if (File.Exists(path + ".lock") && File.GetCreationTime(path + ".lock") > DateTime.Now.AddSeconds(-timeout))
+                throw new Exception("Unable to aqcuire lock");
+            this.fileLock = SimpleFileLock.Create(path + ".lock", TimeSpan.FromSeconds(timeout));
+            this.expiration = DateTime.Now.AddSeconds(timeout);
+        }
+        /// <summary>
+        /// Complete this message, remove from queue.
+        /// </summary>
         public void Complete()
         {
+            if (DateTime.Now >= expiration)
+            {
+                throw new Exception("Lock expired");
+            }
             File.Delete(this.path);
             fileLock.ReleaseLock();
         }
-
+        /// <summary>
+        /// Renew the lock on this message for the specified timeout
+        /// </summary>
+        public void RenewLock()
+        {
+            Lock();
+        }
+        /// <summary>
+        /// Release this message making it available to new Receive() calls.
+        /// </summary>
         public void Release()
         {
             fileLock.ReleaseLock();
         }
-
+        /// <summary>
+        /// Retreive the contained object in this message.
+        /// </summary>
+        /// <returns></returns>
         public T Content()
         {
-            return JsonConvert.DeserializeObject<T>(File.ReadAllText(path));
+            return content;
         }
     }
 }
